Auth Create part 1 & Middlewares
